몹 스폰 확률 double 지원 추가 (최소 0.01)

+++ src/main/java/me/sul/worldmanager/summonmob/mobtype/AutoSummonableMob.java
@@ -4,7 +4,7 @@ import org.bukkit.Location;
 
 public interface AutoSummonableMob {
     void summonMob(Location loc);
-    int getSpawnChance();
+    double getSpawnChance();
     int getMinDistance();
     int getMaxDistance();
 }
