+++ src/main/java/me/sul/worldmanager/summonmob/mobtype/Skeleton.java
@@ -5,13 +5,13 @@ import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.EntityType;
 
 public class Skeleton implements AutoSummonableMob {
-    private final int SPAWN_CHANGE;
+    private final double SPAWN_CHANGE;
     private final int MIN_DISTANCE;
     private final int MAX_DISTANCE;
 
     public Skeleton(FileConfiguration config, String parentNode) {
         parentNode = parentNode + ".skeleton";
-        SPAWN_CHANGE = config.getInt(parentNode + ".spawn-chance");
+        SPAWN_CHANGE = config.getDouble(parentNode + ".spawn-chance");
         MIN_DISTANCE = config.getInt(parentNode + ".min-distance");
         MAX_DISTANCE = config.getInt(parentNode + ".max-distance");
     }
@@ -22,7 +22,7 @@ public class Skeleton implements AutoSummonableMob {
     }
 
     @Override
-    public int getSpawnChance() { return SPAWN_CHANGE; }
+    public double getSpawnChance() { return SPAWN_CHANGE; }
     @Override
     public int getMinDistance() { return MIN_DISTANCE; }
     @Override
